+++ lib/oxidized/input/ssh.rb
@@ -2,17 +2,14 @@ module Oxidized
   require 'net/ssh'
   require 'net/ssh/proxy/command'
   require 'timeout'
-  require 'oxidized/input/cli'
-  class SSH < Input
-    include Input::CLI
+  require_relative 'sshbase'
+
+  class SSH < SSHBase
+    class NoShell < OxidizedError; end
 
     RESCUE_FAIL = {
-      Net::SSH::Disconnect           => :debug,
-      RuntimeError                   => :warn,
-      Net::SSH::AuthenticationFailed => :warn
+      RuntimeError => :warn
     }.freeze
-
-    class NoShell < OxidizedError; end
 
     def self.rescue_fail
       super.merge(RESCUE_FAIL)
@@ -40,10 +37,6 @@ module Oxidized
         end
       end
       connected?
-    end
-
-    def connected?
-      @ssh && (not @ssh.closed?)
     end
 
     def cmd(cmd, expect = node.prompt)
@@ -137,49 +130,6 @@ module Oxidized
           true
         end
       end
-    end
-
-    def make_ssh_opts
-      secure = Oxidized.config.input.ssh.secure?
-      ssh_opts = {
-        number_of_password_prompts:      0,
-        keepalive:                       vars(:ssh_no_keepalive) ? false : true,
-        verify_host_key:                 secure ? :always : :never,
-        append_all_supported_algorithms: true,
-        password:                        @node.auth[:password],
-        timeout:                         @node.timeout,
-        port:                            (vars(:ssh_port) || 22).to_i,
-        forward_agent:                   false
-      }
-
-      auth_methods = vars(:auth_methods) || %w[none publickey password]
-      ssh_opts[:auth_methods] = auth_methods
-      logger.debug "AUTH METHODS::#{auth_methods}"
-
-      ssh_opts[:proxy] = make_ssh_proxy_command(vars(:ssh_proxy), vars(:ssh_proxy_port), secure) if vars(:ssh_proxy)
-
-      ssh_opts[:keys]       = [vars(:ssh_keys)].flatten           if vars(:ssh_keys)
-      ssh_opts[:kex]        = vars(:ssh_kex).split(/,\s*/)        if vars(:ssh_kex)
-      ssh_opts[:encryption] = vars(:ssh_encryption).split(/,\s*/) if vars(:ssh_encryption)
-      ssh_opts[:host_key]   = vars(:ssh_host_key).split(/,\s*/)   if vars(:ssh_host_key)
-      ssh_opts[:hmac]       = vars(:ssh_hmac).split(/,\s*/)       if vars(:ssh_hmac)
-
-      # Use our logger for Net::SSH
-      ssh_logger = SemanticLogger[Net::SSH]
-      ssh_logger.level = Oxidized.config.input.debug? ? :debug : :fatal
-      ssh_opts[:logger] = ssh_logger
-
-      ssh_opts
-    end
-
-    def make_ssh_proxy_command(proxy_host, proxy_port, secure)
-      return nil unless !proxy_host.nil? && !proxy_host.empty?
-
-      proxy_command =  "ssh "
-      proxy_command += "-o StrictHostKeyChecking=no " unless secure
-      proxy_command += "-p #{proxy_port} "            if proxy_port
-      proxy_command += "#{proxy_host} -W [%h]:%p"
-      Net::SSH::Proxy::Command.new(proxy_command)
     end
   end
 end